+++ src/language-markdown/parse/micromark/html-flow-hack.js
@@ -4,7 +4,9 @@ To enforce all html tags parse like `<div>`
 https://github.com/remarkjs/remark-gfm/issues/81
 */
 import { fromMarkdown as originalFromMarkdown } from "mdast-util-from-markdown";
+import { htmlFlow } from "micromark-core-commonmark";
 import { htmlBlockNames, htmlRawNames } from "micromark-util-html-tag-name";
+import { codes } from "micromark-util-symbol";
 import * as assert from "#universal/assert";
 
 const isProduction = process.env.NODE_ENV !== "production";
@@ -48,4 +50,16 @@ if (isProduction) {
   };
 }
 
-export { fromMarkdown };
+// interrupts html-text by html-flow
+function htmlFlowHackSyntax() {
+  return {
+    text: {
+      [codes.lessThan]: {
+        ...htmlFlow,
+        add: "before",
+      },
+    },
+  };
+}
+
+export { fromMarkdown, htmlFlowHackSyntax };
