+++ src/language-markdown/traverse/visitor-keys.evaluate.js
@@ -35,6 +35,7 @@ const visitorKeys = generateReferenceSharedVisitorKeys({
   esComment: [],
   jsx: [],
   math: [],
+  newLineHack: [],
   inlineMath: [],
   tableRow: ["children"],
   listItem: ["children"],
